// rbac-admin/src/pages/admin/accesss/List/data.d.ts
export interface TableListItem {
  _id:string;
  name: string;
  desc: string;
  updatedAt: Date;
  createdAt: Date;
}

export interface TableListPagination {
  total: number;
  pageSize: number;
  current: number;
}

export interface TableListData {
  list: TableListItem[];
  pagination: Partial<TableListPagination>;
}

export interface TableListParams {
  _id?:string;
  name?: string;
  desc?: string;
  updatedAt?: Date;
  createdAt?: Date;
}